requirements.txt
python-gnupg
mechanize
termcolor
tabulate
mock
